Revert "danza kuduro"
This reverts commit 5f392dd6993d1dd101a4022278fea4fd282b5d24.

+++ backend/src/main/kotlin/com/wuji/backend/reports/QuestionsSummaryReport.kt
@@ -62,8 +62,7 @@ object QuestionsSummaryReport : Report {
                         incorrectCount,
                         "%.2f"
                             .format(
-                                correctCount.toDouble() /
-                                    (correctCount + incorrectCount).toDouble() *
+                                correctCount / (correctCount + incorrectCount) *
                                     100f)))
             }
         }
